refactor(sql-editor): extract setConnectionByTab as a composable function
